Move rest-layer-mem out of vendor folder
