+++ examples/demo/main.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/justinas/alice"
 	"github.com/rs/cors"
-	"github.com/rs/rest-layer-mem"
+	"github.com/rs/rest-layer/resource/testing/mem"
 	"github.com/rs/rest-layer/resource"
 	"github.com/rs/rest-layer/rest"
 	"github.com/rs/rest-layer/schema"
