+++ examples/projection/main.go
@@ -12,7 +12,7 @@ import (
 	"time"
 
 	"github.com/justinas/alice"
-	"github.com/rs/rest-layer-mem"
+	"github.com/rs/rest-layer/resource/testing/mem"
 	"github.com/rs/rest-layer/resource"
 	"github.com/rs/rest-layer/rest"
 	"github.com/rs/rest-layer/schema"
