+++ resource/testing/mem/latency.go
