+++ rest/example_test.go
@@ -6,7 +6,7 @@ import (
 	"net/url"
 
 	"github.com/rs/cors"
-	"github.com/rs/rest-layer-mem"
+	"github.com/rs/rest-layer/resource/testing/mem"
 	"github.com/rs/rest-layer/resource"
 	"github.com/rs/rest-layer/rest"
 	"github.com/rs/rest-layer/schema"
