+++ rest/method_item_delete_test.go
@@ -6,8 +6,8 @@ import (
 	"net/url"
 	"testing"
 
-	"github.com/rs/rest-layer-mem"
 	"github.com/rs/rest-layer/resource"
+	"github.com/rs/rest-layer/resource/testing/mem"
 	"github.com/rs/rest-layer/schema"
 	"github.com/rs/rest-layer/schema/query"
 	"github.com/stretchr/testify/assert"
