+++ rest/resource_path_test.go
@@ -3,7 +3,7 @@ package rest
 import (
 	"testing"
 
-	mem "github.com/rs/rest-layer-mem"
+	mem "github.com/rs/rest-layer/resource/testing/mem"
 	"github.com/rs/rest-layer/resource"
 	"github.com/rs/rest-layer/schema"
 	"github.com/stretchr/testify/assert"
